+++ src/input/analysis.rs
@@ -3,6 +3,7 @@
 
 //! Contains the implementation of the main `Analysis` structure and its methods.
 
+use std::fmt;
 use std::fs::read_to_string;
 use std::path::Path;
 
@@ -10,6 +11,7 @@ use derive_builder::Builder;
 use getset::{CopyGetters, Getters};
 use getset::{MutGetters, Setters};
 use groan_rs::files::FileType;
+use serde::de::{self, SeqAccess, Visitor};
 use serde::{Deserialize, Deserializer, Serialize};
 
 use crate::errors::ConfigError;
@@ -95,10 +97,12 @@ pub struct Analysis {
     #[getset(get = "pub")]
     bonds: Option<String>,
 
-    /// Path to a XTC trajectory file containing the trajectory to be analyzed.
-    #[builder(setter(into))]
+    /// Path to an XTC (recommended), TRR, GRO, PDB, NC, DCD, or LAMMPSTRJ containing the trajectory to be analyzed.
+    /// Multiple XTC or TRR trajectories can be provided and these will be concatenated.
     #[getset(get = "pub")]
-    trajectory: String,
+    #[serde(deserialize_with = "deserialize_string_or_vec")]
+    #[builder(setter(custom))]
+    trajectory: Vec<String>,
 
     /// Optional path to an NDX file containing the groups associated with the system.
     #[builder(setter(into, strip_option), default)]
@@ -298,17 +302,84 @@ fn validate_structure_format(file: &str) -> Result<(), ConfigError> {
     }
 }
 
-fn validate_trajectory_format(file: &str) -> Result<(), ConfigError> {
-    match FileType::from_name(file) {
-        FileType::XTC
-        | FileType::TRR
-        | FileType::GRO
-        | FileType::PDB
-        | FileType::NC
-        | FileType::DCD
-        | FileType::LAMMPSTRJ => Ok(()),
-        _ => Err(ConfigError::InvalidTrajectoryFormat(file.to_owned())),
+fn validate_trajectory_format(files: &[String]) -> Result<(), ConfigError> {
+    if files.len() == 0 {
+        return Err(ConfigError::NoTrajectoryFile);
     }
+
+    let mut last: Option<(FileType, &String)> = None;
+    for file in files {
+        let file_type = FileType::from_name(file);
+        match file_type {
+            FileType::XTC | FileType::TRR => (),
+            FileType::GRO | FileType::PDB | FileType::NC | FileType::DCD | FileType::LAMMPSTRJ => {
+                // trajectory concatenation is not supported for these formats
+                if files.len() > 1 {
+                    return Err(ConfigError::TrajCatNotSupported);
+                }
+            }
+            _ => return Err(ConfigError::InvalidTrajectoryFormat(file.to_owned())),
+        }
+
+        // check that all the trajectory files have the same format
+        if let Some((unwrapped_type, unwrapped_name)) = last {
+            if file_type != unwrapped_type {
+                return Err(ConfigError::InconsistentTrajectoryFormat(
+                    file.clone(),
+                    unwrapped_name.clone(),
+                ));
+            }
+
+            last = Some((file_type, file))
+        }
+    }
+
+    Ok(())
+}
+
+pub fn deserialize_string_or_vec<'de, D>(deserializer: D) -> Result<Vec<String>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    struct StringOrVecVisitor;
+
+    impl<'de> Visitor<'de> for StringOrVecVisitor {
+        type Value = Vec<String>;
+
+        fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+            formatter.write_str("a string or a sequence of strings")
+        }
+
+        fn visit_str<E>(self, value: &str) -> Result<Self::Value, E>
+        where
+            E: de::Error,
+        {
+            match glob::glob(value) {
+                Ok(paths) => {
+                    let files: Vec<_> = paths
+                        .filter_map(Result::ok)
+                        .map(|path| path.to_string_lossy().into_owned())
+                        .collect();
+
+                    if files.is_empty() {
+                        Ok(vec![value.to_owned()])
+                    } else {
+                        Ok(files)
+                    }
+                }
+                Err(e) => Err(de::Error::custom(e.to_string())),
+            }
+        }
+
+        fn visit_seq<A>(self, seq: A) -> Result<Self::Value, A::Error>
+        where
+            A: SeqAccess<'de>,
+        {
+            Deserialize::deserialize(de::value::SeqAccessDeserializer::new(seq))
+        }
+    }
+
+    deserializer.deserialize_any(StringOrVecVisitor)
 }
 
 fn deserialize_membrane_normal<'de, D>(deserializer: D) -> Result<MembraneNormal, D::Error>
@@ -467,7 +538,84 @@ impl Analysis {
     }
 }
 
+/// Helper enum for providing input trajectory file(s).
+#[derive(Debug, Clone)]
+pub enum TrajectoryInput {
+    Single(String),
+    Multiple(Vec<String>),
+}
+
+impl TrajectoryInput {
+    fn expand<T>(value: T) -> Self
+    where
+        T: Into<String>,
+    {
+        let string_pattern = value.into();
+        match glob::glob(&string_pattern) {
+            Ok(paths) => {
+                let files: Vec<_> = paths
+                    .filter_map(Result::ok)
+                    .map(|path| path.to_string_lossy().into_owned())
+                    .collect();
+
+                // if there is no match, just use the original pattern
+                if files.is_empty() {
+                    Self::Single(string_pattern)
+                } else {
+                    Self::Multiple(files)
+                }
+            }
+            Err(e) => panic!(
+                "FATAL GORDER ERROR | TrajectoryInput::expand | Trajectory glob pattern is invalid. Error: {}. (This may be your fault.)", e)
+        }
+    }
+}
+
+impl From<&str> for TrajectoryInput {
+    fn from(value: &str) -> Self {
+        Self::expand(value)
+    }
+}
+
+impl From<String> for TrajectoryInput {
+    fn from(value: String) -> Self {
+        Self::expand(value)
+    }
+}
+
+impl From<&String> for TrajectoryInput {
+    fn from(value: &String) -> Self {
+        Self::expand(value)
+    }
+}
+
+impl<T> From<Vec<T>> for TrajectoryInput
+where
+    T: Into<String>,
+{
+    fn from(value: Vec<T>) -> Self {
+        Self::Multiple(value.into_iter().map(|x| x.into()).collect())
+    }
+}
+
+impl<T> From<&[T]> for TrajectoryInput
+where
+    T: Into<String> + Clone,
+{
+    fn from(value: &[T]) -> Self {
+        Self::Multiple(value.iter().cloned().map(|x| x.into()).collect())
+    }
+}
+
 impl AnalysisBuilder {
+    pub fn trajectory(&mut self, trajectory: impl Into<TrajectoryInput>) -> &mut Self {
+        match trajectory.into() {
+            TrajectoryInput::Single(x) => self.trajectory = Some(vec![x]),
+            TrajectoryInput::Multiple(x) => self.trajectory = Some(x),
+        }
+        self
+    }
+
     /// Be silent. Print nothing to the standard output during the analysis.
     #[inline(always)]
     pub fn silent(&mut self) -> &mut Self {
@@ -557,8 +705,8 @@ mod tests_yaml {
     fn analysis_yaml_pass_basic() {
         let analysis = Analysis::from_file("tests/files/inputs/basic.yaml").unwrap();
 
-        assert_eq!(analysis.structure(), "system.tpr");
-        assert_eq!(analysis.trajectory(), "md.xtc");
+        assert_eq!(analysis.structure(), "tests/files/pcpepg.tpr");
+        assert_eq!(analysis.trajectory(), &vec!["tests/files/pcpepg.xtc"]);
         assert!(analysis.index().is_none());
         assert!(analysis.output().is_none());
         assert!(analysis.output_tab().is_none());
@@ -594,8 +742,8 @@ mod tests_yaml {
     fn analysis_yaml_pass_full() {
         let analysis = Analysis::from_file("tests/files/inputs/full.yaml").unwrap();
 
-        assert_eq!(analysis.structure(), "system.tpr");
-        assert_eq!(analysis.trajectory(), "md.xtc");
+        assert_eq!(analysis.structure(), "tests/files/cg.tpr");
+        assert_eq!(analysis.trajectory(), &["tests/files/cg.xtc"]);
         assert_eq!(analysis.index().as_ref().unwrap(), "index.ndx");
         assert_eq!(analysis.output().as_ref().unwrap(), "order.yaml");
         assert_eq!(analysis.output_tab().as_ref().unwrap(), "order.dat");
@@ -1004,8 +1152,8 @@ mod tests_builder {
     #[test]
     fn analysis_builder_pass_basic() {
         let analysis = Analysis::builder()
-            .structure("system.tpr")
-            .trajectory("md.xtc")
+            .structure("tests/files/pcpepg.tpr")
+            .trajectory("tests/files/pcpepg.xtc")
             .analysis_type(AnalysisType::aaorder(
                 "@membrane and element name carbon",
                 "@membrane and element name hydrogen",
@@ -1013,8 +1161,8 @@ mod tests_builder {
             .build()
             .unwrap();
 
-        assert_eq!(analysis.structure(), "system.tpr");
-        assert_eq!(analysis.trajectory(), "md.xtc");
+        assert_eq!(analysis.structure(), "tests/files/pcpepg.tpr");
+        assert_eq!(analysis.trajectory(), &["tests/files/pcpepg.xtc"]);
         assert!(analysis.index().is_none());
         assert!(analysis.output().is_none());
         assert!(analysis.output_tab().is_none());
@@ -1049,8 +1197,8 @@ mod tests_builder {
     #[test]
     fn analysis_builder_pass_full() {
         let analysis = Analysis::builder()
-            .structure("system.tpr")
-            .trajectory("md.xtc")
+            .structure("tests/files/cg.tpr")
+            .trajectory("tests/files/cg.xtc")
             .index("index.ndx")
             .output("order.yaml")
             .output_tab("order.dat")
@@ -1091,8 +1239,8 @@ mod tests_builder {
             .build()
             .unwrap();
 
-        assert_eq!(analysis.structure(), "system.tpr");
-        assert_eq!(analysis.trajectory(), "md.xtc");
+        assert_eq!(analysis.structure(), "tests/files/cg.tpr");
+        assert_eq!(analysis.trajectory(), &["tests/files/cg.xtc"]);
         assert_eq!(analysis.index().as_ref().unwrap(), "index.ndx");
         assert_eq!(analysis.output().as_ref().unwrap(), "order.yaml");
         assert_eq!(analysis.output_tab().as_ref().unwrap(), "order.dat");
@@ -1164,8 +1312,8 @@ mod tests_builder {
     #[test]
     fn analysis_builder_pass_default_ordermap() {
         let analysis = Analysis::builder()
-            .structure("system.tpr")
-            .trajectory("md.xtc")
+            .structure("tests/files/pcpepg.tpr")
+            .trajectory("tests/files/pcpepg.xtc")
             .output("order.yaml")
             .analysis_type(AnalysisType::aaorder(
                 "@membrane and element name carbon",
@@ -1187,8 +1335,8 @@ mod tests_builder {
     #[test]
     fn analysis_builder_pass_dynamic_membrane_normal() {
         let analysis = Analysis::builder()
-            .structure("system.tpr")
-            .trajectory("md.xtc")
+            .structure("tests/files/cg.tpr")
+            .trajectory("tests/files/cg.xtc")
             .output("order.yaml")
             .analysis_type(AnalysisType::cgorder("@membrane"))
             .membrane_normal(DynamicNormal::new("name PO4", 2.5).unwrap())
@@ -1209,8 +1357,8 @@ mod tests_builder {
     #[test]
     fn analysis_builder_pass_membrane_normal_from_file() {
         let analysis = Analysis::builder()
-            .structure("system.tpr")
-            .trajectory("md.xtc")
+            .structure("tests/files/cg.tpr")
+            .trajectory("tests/files/cg.xtc")
             .output("order.yaml")
             .analysis_type(AnalysisType::cgorder("@membrane"))
             .membrane_normal("normals.yaml")
@@ -1245,8 +1393,8 @@ mod tests_builder {
         );
 
         let analysis = Analysis::builder()
-            .structure("system.tpr")
-            .trajectory("md.xtc")
+            .structure("tests/files/cg.tpr")
+            .trajectory("tests/files/cg.xtc")
             .output("order.yaml")
             .analysis_type(AnalysisType::cgorder("@membrane"))
             .membrane_normal(map)
@@ -1276,8 +1424,8 @@ mod tests_builder {
     #[test]
     fn analysis_builder_fail_incomplete() {
         match Analysis::builder()
-            .structure("system.tpr")
-            .trajectory("md.xtc")
+            .structure("tests/files/cg.tpr")
+            .trajectory("tests/files/cg.xtc")
             .output("order.yaml")
             .build()
         {
@@ -1290,8 +1438,8 @@ mod tests_builder {
     #[test]
     fn analysis_builder_fail_zero_step() {
         match Analysis::builder()
-            .structure("system.tpr")
-            .trajectory("md.xtc")
+            .structure("tests/files/pcpepg.tpr")
+            .trajectory("tests/files/pcpepg.xtc")
             .output("order.yaml")
             .analysis_type(AnalysisType::aaorder(
                 "@membrane and element name carbon",
@@ -1309,8 +1457,8 @@ mod tests_builder {
     #[test]
     fn analysis_builder_fail_zero_min_samples() {
         match Analysis::builder()
-            .structure("system.tpr")
-            .trajectory("md.xtc")
+            .structure("tests/files/pcpepg.tpr")
+            .trajectory("tests/files/pcpepg.xtc")
             .output("order.yaml")
             .analysis_type(AnalysisType::aaorder(
                 "@membrane and element name carbon",
@@ -1328,8 +1476,8 @@ mod tests_builder {
     #[test]
     fn analysis_builder_fail_zero_threads() {
         match Analysis::builder()
-            .structure("system.tpr")
-            .trajectory("md.xtc")
+            .structure("tests/files/pcpepg.tpr")
+            .trajectory("tests/files/pcpepg.xtc")
             .output("order.yaml")
             .analysis_type(AnalysisType::aaorder(
                 "@membrane and element name carbon",
@@ -1347,8 +1495,8 @@ mod tests_builder {
     #[test]
     fn analysis_builder_fail_start_higher_than_end() {
         match Analysis::builder()
-            .structure("system.tpr")
-            .trajectory("md.xtc")
+            .structure("tests/files/pcpepg.tpr")
+            .trajectory("tests/files/pcpepg.xtc")
             .output("order.yaml")
             .analysis_type(AnalysisType::aaorder(
                 "@membrane and element name carbon",
@@ -1367,8 +1515,8 @@ mod tests_builder {
     #[test]
     fn analysis_builder_fail_invalid_structure_format() {
         match Analysis::builder()
-            .structure("system.ndx")
-            .trajectory("md.xtc")
+            .structure("tests/files/pcpepg.ndx")
+            .trajectory("tests/files/pcpepg.xtc")
             .output("order.yaml")
             .analysis_type(AnalysisType::aaorder(
                 "@membrane and element name carbon",
@@ -1385,8 +1533,8 @@ mod tests_builder {
     #[test]
     fn analysis_builder_fail_invalid_trajectory_format() {
         match Analysis::builder()
-            .structure("system.tpr")
-            .trajectory("md.xyz")
+            .structure("tests/files/pcpepg.tpr")
+            .trajectory("tests/files/cg_order_leaflets.tab")
             .output("order.yaml")
             .analysis_type(AnalysisType::aaorder(
                 "@membrane and element name carbon",